refactor: more members moved to Landscape
